+++ src/Tools/Base.gd
@@ -27,11 +27,21 @@ func _on_PixelPerfect_toggled(button_pressed : bool):
 func _on_Horizontal_toggled(button_pressed : bool):
 	tool_slot.horizontal_mirror = button_pressed
 	tool_slot.save_config()
+	Global.show_y_symmetry_axis = button_pressed
+	# If the button is not pressed but another button is, keep the symmetry guide visible
+	if !button_pressed and (Tools._slots[BUTTON_LEFT].horizontal_mirror or Tools._slots[BUTTON_RIGHT].horizontal_mirror):
+		Global.show_y_symmetry_axis = true
+	Global.current_project.y_symmetry_axis.visible = Global.show_y_symmetry_axis and Global.show_guides
 
 
 func _on_Vertical_toggled(button_pressed : bool):
 	tool_slot.vertical_mirror = button_pressed
 	tool_slot.save_config()
+	Global.show_x_symmetry_axis = button_pressed
+	# If the button is not pressed but another button is, keep the symmetry guide visible
+	if !button_pressed and (Tools._slots[BUTTON_LEFT].vertical_mirror or Tools._slots[BUTTON_RIGHT].vertical_mirror):
+		Global.show_x_symmetry_axis = true
+	Global.current_project.x_symmetry_axis.visible = Global.show_x_symmetry_axis and Global.show_guides
 
 
 func save_config() -> void: